Integrator builder.
Abstractions to allow integrators with different requirements to be
built with a common interface. Initial implementation uses the previous
Integrator object as both a dispatcher for the current method
implementations as well as for a container of parameters that do not
yet have their own method in the Builder interface.