Framework for make_operation.
Utility to be used to allow easy export of Python classes as compliant
Operation factories. Generates a factory function that produces a valid
Operation instance. Generates the Operation subclass to be used by the
factory. Generates and validates the required Operation interface, which
includes a callable `run` property and an `output` property with
prescribed behaviors.

To do:
* flesh out the Operation behaviors
* map input, output, and callable expressed in utility function call
  to the wrapped class