Initial commit of infrastructure supporting fork.
Acts as a patch for GROMACS master (git ref of branch point noted in
.travis.yml) to serve as a base for Kasson Lab fork.

Feature branches based on the Kasson Lab fork are easily converted to
Gerrit changes based on the represented ref in master.

Not targeted for merging with master. Could serve as a sandbox branch
in Gerrit.

This point in the history is based on commit bec0fa7b9 in master.